+++ src/KSPACE/pppm_dipole.cpp
@@ -2507,7 +2507,7 @@ int PPPMDipole::timing_3d(int n, double &time3d)
 
 double PPPMDipole::memory_usage()
 {
-  double bytes = nmax*3 * sizeof(double);
+  double bytes = (double)nmax*3 * sizeof(double);
 
   int nbrick = (nxhi_out-nxlo_out+1) * (nyhi_out-nylo_out+1) *
     (nzhi_out-nzlo_out+1);
