+++ src/MOLECULE/fix_cmap.cpp
@@ -1440,7 +1440,7 @@ int FixCMAP::unpack_exchange(int nlocal, double *buf)
 double FixCMAP::memory_usage()
 {
   int nmax = atom->nmax;
-  double bytes = nmax * sizeof(int);        // num_crossterm
+  double bytes = (double)nmax * sizeof(int);        // num_crossterm
   bytes += (double)nmax*CMAPMAX * sizeof(int);      // crossterm_type
   bytes += (double)5*nmax*CMAPMAX * sizeof(int);    // crossterm_atom 12345
   bytes += (double)maxcrossterm*6 * sizeof(int);    // crosstermlist
