+++ src/USER-FEP/pair_tip4p_long_soft.cpp
@@ -510,7 +510,7 @@ void *PairTIP4PLongSoft::extract(const char *str, int &dim)
 
 double PairTIP4PLongSoft::memory_usage()
 {
-  double bytes = maxeatom * sizeof(double);
+  double bytes = (double)maxeatom * sizeof(double);
   bytes += (double)maxvatom*6 * sizeof(double);
   bytes += (double)2 * nmax * sizeof(double);
   return bytes;
