+++ src/USER-PHONON/fix_phonon.cpp
@@ -424,7 +424,7 @@ void FixPhonon::end_of_step()
 
 double FixPhonon::memory_usage()
 {
-  double bytes = sizeof(double)*2*mynq
+  double bytes = (double)sizeof(double)*2*mynq
                + sizeof(std::map<int,int>)*2*ngroup
                + sizeof(double)*(ngroup*(3*sysdim+2)+mynpt*fft_dim*2)
                + sizeof(std::complex<double>)*MAX(1,mynq)*fft_dim *(1+2*fft_dim)
