+++ src/USER-SMD/compute_smd_rho.cpp
@@ -101,6 +101,6 @@ void ComputeSMDRho::compute_peratom() {
  ------------------------------------------------------------------------- */
 
 double ComputeSMDRho::memory_usage() {
-        double bytes = nmax * sizeof(double);
+        double bytes = (double)nmax * sizeof(double);
         return bytes;
 }
