+++ src/USER-SMD/compute_smd_tlsph_shape.cpp
@@ -130,6 +130,6 @@ void ComputeSmdTlsphShape::compute_peratom() {
  ------------------------------------------------------------------------- */
 
 double ComputeSmdTlsphShape::memory_usage() {
-        double bytes = size_peratom_cols * nmax * sizeof(double);
+        double bytes = (double)size_peratom_cols * nmax * sizeof(double);
         return bytes;
 }
