+++ src/bond.cpp
@@ -329,7 +329,7 @@ void Bond::write_file(int narg, char **arg)
 
 double Bond::memory_usage()
 {
-  double bytes = comm->nthreads*maxeatom * sizeof(double);
+  double bytes = (double)comm->nthreads*maxeatom * sizeof(double);
   bytes += (double)comm->nthreads*maxvatom*6 * sizeof(double);
   return bytes;
 }
