+++ src/improper_hybrid.cpp
@@ -350,7 +350,7 @@ void ImproperHybrid::read_restart(FILE *fp)
 
 double ImproperHybrid::memory_usage()
 {
-  double bytes = maxeatom * sizeof(double);
+  double bytes = (double)maxeatom * sizeof(double);
   bytes += (double)maxvatom*6 * sizeof(double);
   bytes += (double)maxcvatom*9 * sizeof(double);
   for (int m = 0; m < nstyles; m++) bytes += (double)maximproper[m]*5 * sizeof(int);
